pyproject.toml
[tool.poetry]
name = "rss-reader"
version = "0.1.0"
description = "RSS Reader web application"
authors = ["Guilherme Pereira <guilherme.mp.eng@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.1"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
sqlalchemy = "^2.0.13"
psycopg2-binary = "^2.9.6"
pydantic = "^1.10.7"
alembic = "^1.10.4"
pytest = "^7.3.1"
httpx = "^0.24.0"
email-validator = "^2.0.0.post2"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.2"
celery = "^5.2.7"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
